Second round. Not sure why the first is still showing up here. (#6137)
* Create gradle.yml

* Autorefactor: remove StringBuilder where String does not need to be built incrementally.

* Autorefactor: merge combinations of StringBuilder and concat building.

* Autorefactor: use StringBuilder when building strings in loops.

* Autorefactor: use enhanced for loops when valid.

* Autorefactor: remove use of indexOf to find an item in a collection but not care where it is.

* Autorefactor: replace multi branching if tests with switches.

* Autorefactor: simplify lambdas and inner code.

* Autorefactor: boolean expressions.

* Manual fix to autorefactor incorrectly flagging condition constant

* Autorefactor: (now) unused imports

* Manual fix PMD unused variables

* Restore gradle.yml

* Functional style isPresent

* Autorefactor: Unnecessary semicolons

* Manual: loops that aren't really loops

* Manual: remove booleans "&& false"

* Autorefactor: redundant IFs

* Autorefactor: short circuitable

* Autorefactor: no effect ternaries

* Autorefactor: two way switches

* Autorefactor: duplicated branch endings

* Autorefactor: type boxing/unboxing

* Autorefactor: make inner classes static if enclosing instance doesn't matter

* Autorefactor: construct populated lists instead of constructing empty and adding

* Autorefactor: redundant initializer and assignment

* Manual flag: Strings built but then thrown away?

* Autorefactor: short circuit searching loops

* Autorefactor: inlinings

* Revert some inlinings to pass testing

* Manual tidy for checkstyle/pmd

+++ code/src/java/pcgen/cdom/content/KnownSpellIdentifier.java
@@ -69,7 +69,7 @@ public class KnownSpellIdentifier extends ConcretePrereqObject
 	public KnownSpellIdentifier(CDOMReference<Spell> spellRef, Integer levelLimit)
 	{
 		Objects.requireNonNull(spellRef, "Known Spell Identifier cannot have null spell reference");
-		if (levelLimit != null && levelLimit.intValue() < 0)
+		if (levelLimit != null && levelLimit < 0)
 		{
 			throw new IllegalArgumentException("Known Spell Identifier level limit cannot be negative");
 		}
@@ -166,7 +166,7 @@ public class KnownSpellIdentifier extends ConcretePrereqObject
 	@Override
 	public int hashCode()
 	{
-		return spellLevel == null ? ref.hashCode() : spellLevel.intValue() * ref.hashCode();
+		return spellLevel == null ? ref.hashCode() : spellLevel * ref.hashCode();
 	}
 
 	@Override
