+++ code/src/java/plugin/lsttokens/add/AbilityToken.java
@@ -321,7 +321,7 @@ public class AbilityToken extends AbstractNonEmptyToken<CDOMObject>
 					Integer stackLimit = container.getStackLimit();
 					if (stackLimit != null)
 					{
-						if (stackLimit.intValue() <= 0)
+						if (stackLimit <= 0)
 						{
 							context.addWriteMessage("Stack Limit in " + getFullName() + " must be > 0");
 							return null;
