+++ code/src/java/plugin/lsttokens/equipment/CritrangeToken.java
@@ -48,7 +48,7 @@ public class CritrangeToken implements CDOMPrimaryToken<Equipment>
 		try
 		{
 			Integer cr = Integer.valueOf(value);
-			if (cr.intValue() < 0)
+			if (cr < 0)
 			{
 				return new ParseResult.Fail(getTokenName() + " cannot be < 0");
 			}
