+++ code/src/java/plugin/lsttokens/pcclass/LevelsperfeatToken.java
@@ -88,7 +88,7 @@ public class LevelsperfeatToken extends AbstractTokenWithSeparator<PCClass> impl
 		try
 		{
 			Integer in = Integer.valueOf(numLevels);
-			if (in.intValue() < 0)
+			if (in < 0)
 			{
 				return new ParseResult.Fail(getTokenName() + " must be an integer >= 0");
 			}
@@ -116,7 +116,7 @@ public class LevelsperfeatToken extends AbstractTokenWithSeparator<PCClass> impl
 			}
 			return null;
 		}
-		if (lpf.intValue() < 0)
+		if (lpf < 0)
 		{
 			context.addWriteMessage(getTokenName() + " must be an integer >= 0");
 			return null;
