+++ code/src/java/plugin/primitive/spell/AllToken.java
@@ -55,7 +55,7 @@ public class AllToken extends AbstractRestrictedSpellPrimitive
 		{
 			for (Integer level : levelInfo.getListFor(spellList))
 			{
-				if (allow(pc, level.intValue(), "ANY", spell, null))
+				if (allow(pc, level, "ANY", spell, null))
 				{
 					return true;
 				}
